+++ Sources/TuistEnvKit/Installer/BuildCopier.swift
@@ -18,24 +18,14 @@ class BuildCopier: BuildCopying {
         "libProjectDescription.dylib",
     ]
 
-    // MARK: - Attributes
-
-    private let system: Systeming
-
-    // MARK: - Init
-
-    init(system: Systeming = System()) {
-        self.system = system
-    }
-
     func copy(from: AbsolutePath, to: AbsolutePath) throws {
         try BuildCopier.files.forEach { file in
             let filePath = from.appending(component: file)
             let toPath = to.appending(component: file)
             if !FileHandler.shared.exists(filePath) { return }
-            try system.run("/bin/cp", "-rf", filePath.pathString, toPath.pathString)
+            try System.shared.run("/bin/cp", "-rf", filePath.pathString, toPath.pathString)
             if file == "tuist" {
-                try system.run("/bin/chmod", "+x", toPath.pathString)
+                try System.shared.run("/bin/chmod", "+x", toPath.pathString)
             }
         }
     }